fix(resourceEditor): close review-found correctness gaps
Three issues from a thorough multi-reviewer pass on the branch:

1. `ResourceEditorHost` constructed `useResourceEditor` unconditionally
   with `extension: () => resolvedExtension.value!`. When no extension
   matched (e.g. a resource the registry can't handle), the composable
   read `unref(extensionRef).appId` and crashed before the template
   could render the `#no-editor` slot. Split the host into an outer
   shell that gates on `resolvedExtension` and an inner
   `ResourceEditorMount` that runs the composable only once an
   extension is resolved. Mount stays internal to the package.

2. `web-app-epub-reader` wrapped its App in `defineAsyncComponent` to
   defer the epubjs bundle. The wrapper has no `props`/`emits` until
   the chunk resolves, so the composable's capability introspection
   (`hasProp('currentContent')`) returned `false` and content was
   never loaded. App.vue is now statically imported and the epubjs
   dependency itself is dynamically imported inside App.vue's setup,
   keeping the heavy lib deferred while exposing the component's
   props/emits to the host. Test helper exports `flushPromises` so
   the existing spec can wait for the dynamic import.

3. `useResourceEditor`'s `fileSizeLimit` modal was dispatched whenever
   `resource` or `space` changed. Preview's photo-roll re-fires that
   watcher on every active-file swap, so the modal popped on every
   click. Track the last resource id we prompted for and only
   re-dispatch on actual transitions.
